Between

+++ src/main/java/ru/skypro/homework/dto/AdsDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class AdsDTO {
-    private Integer author;
+    private Long author;
     private String image;
     private Integer pk;
     private Integer price;
